+++ heat/tests/test_get_grid_z.f90
@@ -8,10 +8,10 @@ program test_get_grid_z
 
   integer, parameter :: grid_id = 1
   integer, parameter :: nz = 1
-  real, parameter, dimension(nz) :: expected_z = (/ 0.0 /)
+  double precision, parameter, dimension(nz) :: expected_z = (/ 0.0 /)
 
   type (bmi_heat) :: m
-  real, dimension(nz) :: grid_z
+  double precision, dimension(nz) :: grid_z
   integer :: i
 
   status = m%initialize(config_file)
